refactor: moving utils to their components and url validation
