+++ src/editors/sharedComponents/InsertLinkModal/blockTypes.js
@@ -1,0 +1,7 @@
+const blockTypes = {
+  section: 'chapter',
+  subsection: 'sequential',
+  unit: 'vertical',
+};
+
+export default blockTypes;
