+++ src/editors/sharedComponents/InsertLinkModal/utils.test.js
@@ -1,10 +1,6 @@
 import {
   addPathToBlocks,
   formatBlocks,
-  getSectionsList,
-  getChildrenFromList,
-  filterBlocksByText,
-  formatBlockPath,
   isValidURL,
 } from './utils';
 
@@ -65,7 +61,6 @@ describe('utils', () => {
       expect(nestedBlock2.parentId).toBe(rootBlockId);
     });
   });
-
   describe('formatBlocks', () => {
     const mockBlocks = {
       blockRoot: {
@@ -128,233 +123,6 @@ describe('utils', () => {
       expect(formattedBlocks.block1.displayName).toEqual('Block children 1');
     });
   });
-
-  describe('getSectionsList function', () => {
-    test('returns an empty array for an empty blocks object', () => {
-      const result = getSectionsList({});
-      expect(result).toEqual([]);
-    });
-
-    test('returns an empty array if there are no sections in the blocks object', () => {
-      const blocks = {
-        block1: {
-          id: 'block1',
-          type: 'unit',
-        },
-        block2: {
-          id: 'block2',
-          type: 'vertical',
-        },
-      };
-      const result = getSectionsList(blocks);
-      expect(result).toEqual([]);
-    });
-
-    test('returns an array containing sections from the blocks object', () => {
-      const blocks = {
-        section1: {
-          id: 'section1',
-          type: 'chapter',
-        },
-        block1: {
-          id: 'block1',
-          type: 'unit',
-        },
-        section2: {
-          id: 'section2',
-          type: 'chapter',
-        },
-        block2: {
-          id: 'block2',
-          type: 'vertical',
-        },
-      };
-      const result = getSectionsList(blocks);
-      const expected = [
-        {
-          id: 'section1',
-          type: 'chapter',
-        },
-        {
-          id: 'section2',
-          type: 'chapter',
-        },
-      ];
-      expect(result).toEqual(expected);
-    });
-  });
-
-  describe('getChildrenFromList function', () => {
-    test('returns an empty array when blockSelected has no children', () => {
-      const blocks = {
-        parentBlock: {
-          id: 'parentBlock',
-        },
-      };
-
-      const selectedBlock = blocks.parentBlock;
-      const childrenList = getChildrenFromList(selectedBlock, blocks);
-
-      expect(childrenList).toEqual([]);
-    });
-
-    test('returns an array of child blocks when blockSelected has children', () => {
-      const blocks = {
-        parentBlock: {
-          id: 'parentBlock',
-          children: ['child1', 'child2'],
-        },
-        child1: {
-          id: 'child1',
-        },
-        child2: {
-          id: 'child2',
-        },
-      };
-
-      const selectedBlock = blocks.parentBlock;
-      const childrenList = getChildrenFromList(selectedBlock, blocks);
-
-      expect(childrenList).toHaveLength(2);
-      expect(childrenList).toContainEqual(blocks.child1);
-      expect(childrenList).toContainEqual(blocks.child2);
-    });
-
-    test('returns an empty array when blockSelected.children is undefined', () => {
-      const blocks = {
-        parentBlock: {
-          id: 'parentBlock',
-          children: undefined,
-        },
-      };
-
-      const selectedBlock = blocks.parentBlock;
-      const childrenList = getChildrenFromList(selectedBlock, blocks);
-
-      expect(childrenList).toEqual([]);
-    });
-
-    test('returns an empty array when blockSelected.children is an empty array', () => {
-      const blocks = {
-        parentBlock: {
-          id: 'parentBlock',
-          children: [],
-        },
-      };
-
-      const selectedBlock = blocks.parentBlock;
-      const childrenList = getChildrenFromList(selectedBlock, blocks);
-
-      expect(childrenList).toEqual([]);
-    });
-  });
-
-  describe('filterBlocksByText function', () => {
-    const testBlocks = {
-      block1: {
-        id: 'block1',
-        path: 'Root / Child 1',
-      },
-      block2: {
-        id: 'block2',
-        path: 'Root / Child 2',
-      },
-      block3: {
-        id: 'block3',
-        path: 'Another / Block',
-      },
-    };
-
-    test('returns an empty object when searchText is empty', () => {
-      const searchText = '';
-      const filteredBlocks = filterBlocksByText(searchText, testBlocks);
-      expect(filteredBlocks).toEqual({});
-    });
-
-    test('filters blocks based on case-insensitive searchText', () => {
-      const searchText = 'child';
-      const filteredBlocks = filterBlocksByText(searchText, testBlocks);
-      expect(filteredBlocks).toEqual({
-        block1: {
-          id: 'block1',
-          path: 'Root / Child 1',
-        },
-        block2: {
-          id: 'block2',
-          path: 'Root / Child 2',
-        },
-      });
-    });
-
-    test('returns an empty object when no blocks match searchText', () => {
-      const searchText = 'nonexistent';
-      const filteredBlocks = filterBlocksByText(searchText, testBlocks);
-      expect(filteredBlocks).toEqual({});
-    });
-
-    test('filters blocks with partial matches in path', () => {
-      const searchText = 'root';
-      const filteredBlocks = filterBlocksByText(searchText, testBlocks);
-      expect(filteredBlocks).toEqual({
-        block1: {
-          id: 'block1',
-          path: 'Root / Child 1',
-        },
-        block2: {
-          id: 'block2',
-          path: 'Root / Child 2',
-        },
-      });
-    });
-  });
-
-  describe('formatBlockPath function', () => {
-    test('formats a simple path with title and subtitle', () => {
-      const path = 'Root / Child 1 / Grandchild';
-      const formattedPath = formatBlockPath(path);
-      expect(formattedPath).toEqual({
-        title: 'Grandchild',
-        subTitle: 'Root / Child 1',
-      });
-    });
-
-    test('handles an empty title by using the previous part as title', () => {
-      const path = 'Root / Child 1 / ';
-      const formattedPath = formatBlockPath(path);
-      expect(formattedPath).toEqual({
-        title: 'Child 1',
-        subTitle: 'Root / Child 1',
-      });
-    });
-
-    test('handles an empty path by returning an empty title and subtitle', () => {
-      const path = '';
-      const formattedPath = formatBlockPath(path);
-      expect(formattedPath).toEqual({
-        title: '',
-        subTitle: '',
-      });
-    });
-
-    test('handles whitespace in the title by using the previous part as title', () => {
-      const path = 'Root / Child 1 /   ';
-      const formattedPath = formatBlockPath(path);
-      expect(formattedPath).toEqual({
-        title: 'Child 1',
-        subTitle: 'Root / Child 1',
-      });
-    });
-
-    test('handles a path with only one part by using it as the title', () => {
-      const path = 'SinglePart';
-      const formattedPath = formatBlockPath(path);
-      expect(formattedPath).toEqual({
-        title: 'SinglePart',
-        subTitle: '',
-      });
-    });
-  });
-
   describe('isValidURL function', () => {
     test('returns true for a valid HTTP URL', () => {
       const validHTTPUrl = 'http://www.example.com';
